Update README.md
Network Utilities

Determine which command/utility is needed to accomplish each part and perform the command.

Part 1:

Use the command to view the configuration details, including the IP address, MAC address, and other details about the network interfaces on your Linux system.
Take a screenshot of both the command you entered and the output. Save the screenshot as screenshot_1.
Part 2:

Use the command to test the network connection between your Linux machine and the ASU web server located at asu.edu. Ensure that when you run the command, the command only sends 5 messages to test the connection. 
Take a screenshot of both the command you entered and the output. Save the screenshot as screenshot_2.
Part 3:

Use the command to view the hops between your Linux system and the ASU web server located at asu.edu.
Take a screenshot of both the command you entered and the output. Save the screenshot as screenshot_3.
Part 4:

Use the command to view the kernel's IP routing table.
Take a screenshot of both the command you entered and the output. Save the screenshot as screenshot_4.
Part 5:

Use the command to view the summary of the network statistics for your system.
Take a screenshot of both the command you entered and the output. Save the screenshot as screenshot_5.
Part 6

Do some research to answer to perform this task.
Use the command to find out who the registrant is, along with the administrative contact, technical contact, and date that the domain was first activated for asu.edu.
Take a screenshot of both the command you entered and the output. Save the screenshot as screenshot_6.
You may need to reduce the font size of the terminal or feel free to use multiple screenshots and name them screenshot_6a, screenshot_6b, etc